Use opacity to communicate what block is being dragged.
Previous to this refactor, it used to be a div inside the block wrapper
that would become visible when the block was being dragged. That
div would appear as grey. By refactoring the drag handler to the
breadcrumb we no longer have that div, so we seek to use
a different technique to communicate the same behavior.

I also think is nicer to show the original content.